Introduced support for authentication mechanism
This commit implements an authentication mechanism based on strategies.
There's also a basic strategy ready-to-use.

This mechanism also provides the user the capability to define resource
scopes by adding the Scope attribute to the desired route.

Please note that even though it restricts the access to the specific
resource, it won't stop the user from consuming a license for about 10
seconds.

- Added support for resource scopes

IMPROVEMENTS

- Improved http status resolution
- Added support for authentication using strategies