Improved dev experience for new strategies
This commit removes the flags `Next` and `Required`, instead introducing
a definitive flag called `Verified`. Now strategies must set this flag
to 1 (true) along with a $$$OK return if they want to pass the request.

Note that the manager will automatically skip to the next strategy if this
flag is not set to 1 but still returning $$$OK. Additionaly, as this is the
last one, its challenge will be sent to the client.

Finally, irregular errors must override the $$$OK with their respective error
statuses. This will cause the manager to bail out immediately displaying their
messages, but the challenge won't be sent.